Define build_cmd_split() to reduce code duplication
This is a first step to unify the code of save_part_by_*() functions and
have better error handling.

build_cmd_split define a variable to avoid to spawn a subshell, which
would be slower. The variable encapsulates Use a variable to encapsulate
the command to be used to process the compression's output.

Signed-off-by: Patrick Rouleau <pfrouleau@gmail.com>